Change of account balance replenish
First the counterparty is replenished, then the money is transferred to the agency.
